etl/airflow/requirements.txt
# Dependencies baked into the custom Airflow image for DAGs
vnstock
vnstock3
pandas
pyarrow
mysql-connector-python
openpyxl
beautifulsoup4
seaborn
SQLAlchemy
apache-airflow-providers-postgres
psycopg2-binary
yfinance
